+++ src/core/context.ts
@@ -258,6 +258,7 @@ import {
     createStarPolygon,
 } from "../math/polygongeneration";
 import { insertionSort } from "../math/sort";
+import { makeQuadtree, Quadtree } from "../math/spatial/quadtree";
 import { Vec2 } from "../math/Vec2";
 import { BlendMode, type KAPLAYPlugin } from "../types";
 import {
@@ -535,6 +536,8 @@ export const createContext = (
         Mat4,
         Mat23,
         Quad,
+        Quadtree,
+        makeQuadtree,
         RNG,
         Rule,
         RuleSystem,