+++ src/ecs/systems/createCollisionSystem.ts
@@ -1,15 +1,36 @@
 import { onAdd, onDestroy, onUnuse, onUse } from "../../events/globalEvents";
 import { onSceneLeave } from "../../game/scenes";
+import { height, width } from "../../gfx/stack";
 import { gjkShapeIntersection } from "../../math/gjk";
+import { Rect, vec2 } from "../../math/math";
 import { minkowskiRectShapeIntersection } from "../../math/minkowski";
 import { satShapeIntersection } from "../../math/sat";
-import { SweepAndPrune } from "../../math/spatial/sweepandprune";
+import { Quadtree } from "../../math/spatial/quadtree";
+import {
+    SweepAndPruneHorizontal,
+    SweepAndPruneVertical,
+} from "../../math/spatial/sweepandprune";
 import { _k } from "../../shared";
 import type { GameObj } from "../../types";
 import { type AreaComp, usesArea } from "../components/physics/area";
 import { Collision } from "./Collision";
 
-export const createCollisionSystem = ({ narrow = "gjk" } = {}) => {
+export type BroadPhaseType = "sap" | "sapv" | "quadtree";
+export type NarrowPhaseType = "gjk" | "sat" | "box";
+
+export const createCollisionSystem = (
+    { broad = "sap", narrow = "gjk" }: {
+        broad?: BroadPhaseType;
+        narrow?: NarrowPhaseType;
+    } = {},
+) => {
+    const broadPhaseIntersection = broad === "sap"
+        ? new SweepAndPruneHorizontal()
+        : broad === "sapv"
+        ? new SweepAndPruneVertical()
+        : broad === "quadtree"
+        ? new Quadtree(new Rect(vec2(0, 0), width(), height()), 256, 8)
+        : new SweepAndPruneHorizontal();
     const narrowPhaseIntersection = narrow === "gjk"
         ? gjkShapeIntersection
         : narrow === "sat"
@@ -49,48 +70,47 @@ export const createCollisionSystem = ({ narrow = "gjk" } = {}) => {
         return true;
     }
 
-    const sap = new SweepAndPrune();
-    let sapInit = false;
+    let broadInit = false;
 
     function broadPhase() {
         if (!usesArea()) {
             return;
         }
 
-        if (!sapInit) {
-            sapInit = true;
+        if (!broadInit) {
+            broadInit = true;
             onAdd(obj => {
                 if (obj.has("area")) {
-                    sap.add(obj as GameObj<AreaComp>);
+                    broadPhaseIntersection.add(obj as GameObj<AreaComp>);
                 }
             });
             onDestroy(obj => {
-                sap.remove(obj as GameObj<AreaComp>);
+                broadPhaseIntersection.remove(obj as GameObj<AreaComp>);
             });
             onUse((obj, id) => {
                 if (id === "area") {
-                    sap.add(obj as GameObj<AreaComp>);
+                    broadPhaseIntersection.add(obj as GameObj<AreaComp>);
                 }
             });
             onUnuse((obj, id) => {
                 if (id === "area") {
-                    sap.remove(obj as GameObj<AreaComp>);
+                    broadPhaseIntersection.remove(obj as GameObj<AreaComp>);
                 }
             });
             onSceneLeave(scene => {
-                sapInit = false;
-                sap.clear();
+                broadInit = false;
+                broadPhaseIntersection.clear();
             });
 
             for (const obj of _k.game.root.get("*", { recursive: true })) {
                 if (obj.has("area")) {
-                    sap.add(obj as GameObj<AreaComp>);
+                    broadPhaseIntersection.add(obj as GameObj<AreaComp>);
                 }
             }
         }
 
-        sap.update();
-        for (const [obj1, obj2] of sap) {
+        broadPhaseIntersection.update();
+        for (const [obj1, obj2] of broadPhaseIntersection) {
             narrowPhase(obj1, obj2);
         }
     }